32bitのウォーニングを削除

+++ core/common/cgi.cpp
@@ -377,7 +377,7 @@ std::string unescape_html(const std::string& input)
             if (s[0] != '#')
                 return false;
 
-            for (int i = 1; i < s.size(); i++)
+            for (int i = 1; i < (int)s.size(); i++)
             {
                 if (!(s[i] >= '0' && s[i] <= '9'))
                     return false;
@@ -394,7 +394,7 @@ std::string unescape_html(const std::string& input)
             if (s[0] != '#' || !(s[1] == 'x' || s[1] == 'X'))
                 return false;
 
-            for (int i = 2; i < s.size(); i++)
+            for (int i = 2; i < (int)s.size(); i++)
             {
                 if (!((s[i] >= '0' && s[i] <= '9') ||
                       (s[i] >= 'a' && s[i] <= 'f') ||
