+++ core/common/chandir.cpp
@@ -200,7 +200,7 @@ bool ChannelDirectory::writeChannelVariable(Stream& out, const String& varName, 
 {
     CriticalSection cs(m_lock);
 
-    if (!(index >= 0 && index < m_channels.size()))
+    if (!(index >= 0 && index < (int)m_channels.size()))
         return false;
 
     char buf[1024];
@@ -248,7 +248,7 @@ bool ChannelDirectory::writeFeedVariable(Stream& out, const String& varName, int
 {
     CriticalSection cs(m_lock);
 
-    if (!(index >= 0 && index < m_feeds.size())) {
+    if (!(index >= 0 && index < (int)m_feeds.size())) {
         // empty string
         return true;
     }
@@ -364,7 +364,7 @@ void ChannelDirectory::clearFeeds()
 void ChannelDirectory::setFeedPublic(int index, bool isPublic)
 {
     CriticalSection cs(m_lock);
-    if (index < m_feeds.size())
+    if (index < (int)m_feeds.size())
         m_feeds[index].isPublic = isPublic;
     else
         LOG_DEBUG("setFeedPublic: index %d out of range", index);
